backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
sentence-transformers
numpy
groq